Handling of raw lines with semicolons.
The cursor position of the next newline (pos2) is rolled back to the first semicolon found. To this end a change in next_line() prevents a strip operation from acting on the output raw line so that the index of the semicolon char in the raw line correctly maps into its position in the stream. This shouldn't be a problem since next_fortran_line() already gets rid of spaces except maybe trailing ones that however should be harmless.