Added support for functions parsing.
The parse_subroutine function has been extended to parse functions too. It has been renamed to "parse_routine" as a hint it's not only designed for subroutines. When skipping nested subroutines/functions a stack is used instead of the opened subroutines counter to deal with both inner subroutines and functions.
Minor change: the doxygen interpreter has been extended to consider "retval" and "author" tags too.