Misc fixes.
The parse_use_statement function has been adapted to handle only-lists with entities in the form "local_name => mod_name".
At the beginning of the parse_doxygen function a stream advancing statement is necessary for the subsequent backward scan to work properly.
Handling of interface-blocks with no generic specification (generic name).
Fixed a bug in function prev_raw_line: the line starting cursor index should be increased by one otherwise the first char of the returned string would be a newline. This bug was somehow hidden by a subsequent strip() on that string.